[feat]: Migration to Redis List from RabbitMQ
