+++ util/enqueue.go
@@ -6,8 +6,6 @@ import (
 	"fmt"
 	"os"
 	"time"
-
-	amqp "github.com/rabbitmq/amqp091-go"
 )
 
 func EnqueueRunRequest(ctx context.Context, runID string, fileName string, extension string) error {
@@ -21,39 +19,11 @@ func EnqueueRunRequest(ctx context.Context, runID string, fileName string, exten
 		Timestamp time.Time `json:"timestamp"`
 	}
 
-	// Get RabbitMQ connection string from environment variable or use default
-	rabbitMQURL := os.Getenv("RABBITMQ_URL")
-	if rabbitMQURL == "" {
-		rabbitMQURL = "amqp://guest:guest@localhost:5672/"
-	}
-
-	// Connect to RabbitMQ server
-	conn, err := amqp.Dial(rabbitMQURL)
-	if err != nil {
-		logger.Error(fmt.Sprintf("Failed to connect to RabbitMQ: %v", err))
-	}
-	defer conn.Close()
-
-	// Create a channel
-	ch, err := conn.Channel()
-	if err != nil {
-		logger.Error(fmt.Sprintf("Failed to open a channel: %v", err))
-	}
-	defer ch.Close()
-
 	// Declare a queue
-	queueName := "task_queue"
-	q, err := ch.QueueDeclare(
-		queueName, // name
-		true,      // durable
-		false,     // delete when unused
-		false,     // exclusive
-		false,     // no-wait
-		nil,       // arguments
-	)
-
-	if err != nil {
-		logger.Error(fmt.Sprintf("Failed to declare a queue: %v", err))
+	queueName := os.Getenv("REDIS_QUEUE_NAME")
+	if queueName == "" {
+		queueName = "task_queue"
+		logger.Warn(fmt.Sprintf("REDIS_QUEUE_NAME not set, using default: %s", queueName))
 	}
 
 	// Create a new message
@@ -68,20 +38,11 @@ func EnqueueRunRequest(ctx context.Context, runID string, fileName string, exten
 	body, err := json.Marshal(msg)
 	if err != nil {
 		logger.Error(fmt.Sprintf("Error marshaling message: %v", err))
+		return err
 	}
 
-	// Publish message
-	err = ch.PublishWithContext(
-		ctx,
-		"",     // exchange
-		q.Name, // routing key
-		false,  // mandatory
-		false,  // immediate
-		amqp.Publishing{
-			DeliveryMode: amqp.Persistent,
-			Body:         body,
-		},
-	)
+	// Push message to Redis List (LPUSH = enqueue at head)
+	err = RedisClient.LPush(ctx, queueName, string(body)).Err();
 
 	if err != nil {
 		logger.Error(fmt.Sprintf("Failed to publish message: %v", err))
